Make overzealous PBC error into careful assertion
If we're on a distributed mesh, it's possible for our ghost elements to
have periodic boundary conditions with remote periodic neighbors.  If
we're in such a state and we try to query those neighbors (as can happen
in our adaptive refinement code), we want to get remote_elem back, not
throw an error.

Thanks to @fdkong for catching this.